Move Chooser package-change shortcut reset to list adapter
The shortcut-resetting logic was implemented inline in ChooserActivity's
handler, and this CL moves the v2 implementation to a new callback in `ChooserListAdapter`. The v1 implementation is left as-is, so the design change is a no-op.

This makes the activity-side handling more generic so it can be
moved to the pager-adapter (go/chooser-ntab-refactoring), and it fixes
a minor bug (b/315460869).

This is based on "snapshot 4" and "5" of ag/25335069 (for more context, see that CL or go/chooser-ntab-refactoring).

Bug: 310211468, 315460869
Test: `IntentResolver-tests-{unit,activity,integration}`.
Change-Id: I29e10c801db553edc8132aaf8372140b4c35740e