Favor 'profiles' over 'user handles'/'page numbers'
(Or at least, a first couple changes in that direction.)

Historically the pager-adapter 'profiles' and 'page numbers' have been
equivalent, but if we're trying to refer to a specific profile, we
shouldn't have to know its mapped page number.

These changes are as prototyped in ag/25335069 and described in
go/chooser-ntab-refactoring, in particular replicating the changes of
"snapshot 16" and "snapshot 18." There are more changes planned in
this general direction, but these first two happened to be
dependencies of unrelated work in the same overall refactoring effort,
so I wanted to get them in first.

The first snapshot just switches over a few obvious cases. The second
starts to introduce some distinction in the pager-adapter between
"profiles" and "page numbers," which had historically been equivalent.
There are no behavior changes so far, but this sets up forward-looking
"internal APIs" (helper methods) that we'll go on to use in subsequent
CLs to adapt the remaining usages. See the prototype for more context.

Bug: 310211468
Test: `ResolverActivityTest`/activity tests/presubmit
Change-Id: I226e3801da4514f495a1781a98fa5b4630506163