Clean up indexing/lookup methods in pager adapter
This is originally from snapshot 33 in ag/25335069:

"consolidate and rename-more-explicitly the assorted index
conversion/lookup methods in `MultiProfilePagerAdapter`.

Keeping these methods together (and implemented in terms of each
other where possible) makes it easier to audit our logic as we
prepare for 3+ tabs. Particularly, this establishes a 'source of
truth' for the aspect of our business logic that assigns the clone
profile as a sort of 'secondary handle' for the personal tab."

Bug: 310211468
Test: `IntentResolver-tests-actvity`, `ResolverActivityTest`
Change-Id: I5dda5909318d9741927f3e2657cb4aefdfab6c6b