Removes TargetDataLoader from adapter factory method chain
TargetDataLoader is removed from ActivityLogic since it is now
only used via direct constructor call by ResolverActivity, and
injected directly to a field of ChooserActivity.

Since it is only needed to pass to the ChooserListAdapter
or ResolverListAdapter constructor, this removes the parameter
from being forwarded through a sequence of functions and instead
references the injected parameter directly when calling the
contructor. This allows smoother transition to an assisted-inject
factory.

Bug: 300157408
Test: atest IntentResolver-tests-activity:com.android.intentresolver.v2
Change-Id: I18b230eaf97bf8e26e23e2e1cd1372e7d078520f