Rename CallerInfo -> ActivityLaunch and inject
Rename CallerInfo to ActivityLaunch and renames some existing
properties as appropriate for clarity.

Adds the [Intent] from the activity. This completes the set of
info that is needed as inputs from the caller, containing the
extras with all request parameters.

Updates readChooserRequest require only an ActivityLaunchInfo
instance.

Injects ActivityLaunch into usage sites. This removes the direct
link of reading acitivty.intent, and will allow writing test code
which operates directly on inputs without starting an activity.

Introduces an extension method to minimize duplicated code in the
activities: CreationExtras.addDefaultArgs:

Bug: 300157408
Test: atest IntentResolver-tests-activity
Test: atest IntentResolver-tests-unit:ActivityLaunchTest
Change-Id: Ie132cb3d61e139e03316063186c3ad79d2c488ef