ChooserAdapter: fix NPE for non-work profiles
For some non-work profiles (Communal, and maybe Private), the
getListAdapterForUserHandle is returning null. The method is indeed
nullable. But we run getCount on it without doing a null-check,
causing crashes for such profiles.

This previously couldn't have happened, since shouldShowTabs() returns
false in these cases, and the evaluation was skipped; however,
ag/I8273cf365a1e00b1acff4030086f1a044ad7531f moved that check to later,
so we are hitting this problem.

Bug: 322523699
Bug: 320369524
Bug: 320615143
Bug: 322497953

Test: atest IntentResolver-tests-unit
Change-Id: I82303d6a6a1be1066f5c3e05c53f6a70d486bf5a