Shareousel selection tracker component
A building block for the payload toggling functionality. A component to
track items selection and their relative order. The order is specified
by the set of the initially selected items and can be overriden by the
order on items in the cursor (in case the cursor and the set of initilly
selected items is not in sync, which is not expected).

Bug: 302691505
Test: IntentResolver-tests-unit
Change-Id: I092d9b678c7fbcdc8303e04de0cacbb9d125fa5f