Creates ChooserHelper, a peer class for code cleanup
This is a peer class to ChooserActivity to incrementally assume the
tasks of initialization. Code moved through this class will have
carefully controlled rules (documented in the class) for control and
data flow to prevent additional tangles.

This commit makes only a single control flow change, forwarding the
call to ChooserActivity#init through this class. This makes no change
to functionality yet, but provides a hook for following CLs.

Bug: 309960444
Test: atest com.android.intentresolver
Change-Id: I4d895adb00a09a9d18117639b2d85e3fe880e067