Fix CreationExtras.addDefaultArgs when args absent
addDefaultArgs grabs the existing Bundle from CreationExtras
and adds in the additional values. If default args aren't
present (whenever the initial intent did not have extras),
the updated bundle was never returned within the original
instance since it was relying on modifying the Bundle
in-place.

This corrects the issue by explicitly returning a new instance
containing the updated `DEFAULT_ARGS` Bundle.

Test: atest com.android.intentresolver.v2.ext.CreationExtrasExtTest
Change-Id: I5a7c61baaaeefa2878b4041d809c348bf5ac70c0